# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "user_action_required": {
            "default": false,
            "title": "User Action Required",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ApiResponse": {
        "$defs": {
          "ApiError": {
            "properties": {
              "code": {
                "title": "Code",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "title": "Details",
                "type": "object"
              },
              "message": {
                "title": "Message",
                "type": "string"
              },
              "request_id": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "default": null,
                "title": "Request Id"
              },
              "retryable": {
                "default": false,
                "title": "Retryable",
                "type": "boolean"
              },
              "user_action_required": {
                "default": false,
                "title": "User Action Required",
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "title": "ApiError",
            "type": "object"
          }
        },
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "ApiResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualScheduleTriggerRequest": {
        "properties": {
          "now": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now"
          }
        },
        "title": "ManualScheduleTriggerRequest",
        "type": "object"
      },
...
      "ScheduleTickRequest": {
        "properties": {
          "include_disabled": {
            "default": false,
            "title": "Include Disabled",
            "type": "boolean"
          },
          "now": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now"
          }
        },
        "title": "ScheduleTickRequest",
        "type": "object"
      },
      "ScheduleUpsertRequest": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "interval_seconds": {
            "default": 86400,
            "minimum": 1.0,
            "title": "Interval Seconds",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "payload_template": {
            "additionalProperties": true,
            "title": "Payload Template",
            "type": "object"
          },
          "queue_name": {
            "default": "news:queue:memory",
            "minLength": 1,
            "title": "Queue Name",
            "type": "string"
          },
          "run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run At"
          },
          "schedule_id": {
            "minLength": 1,
            "title": "Schedule Id",
            "type": "string"
          },
          "task_type": {
            "minLength": 1,
            "title": "Task Type",
            "type": "string"
          },
          "trigger_type": {
            "default": "interval",
            "enum": [
              "interval",
              "manual"
            ],
            "title": "Trigger Type",
            "type": "string"
          }
        },
        "required": [
          "schedule_id",
          "name",
          "task_type"
        ],
        "title": "ScheduleUpsertRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NewsRoom API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/schedules": {
      "get": {
        "operationId": "schedules_list",
        "parameters": [
          {
            "in": "query",
            "name": "include_disabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Disabled",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Schedules",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "schedules_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Schedule",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/schedules/tick": {
      "post": {
        "operationId": "schedules_tick",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ScheduleTickRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tick Schedules",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/schedules/{schedule_id}/trigger": {
      "post": {
        "operationId": "schedules_trigger",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ManualScheduleTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Schedule",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
...
}